+++ core/utils/utils.h
@@ -9,7 +9,7 @@ namespace utils
 
 point_type random_point();
 
-long double distance(const point_type& p1, const point_type& p2);
+double distance(const point_type& p1, const point_type& p2);
 
 void dump_adjacency_list(const network& net, const std::string& file_name);
